Update pubdir.h
wydaje mi się że atoi() tylko działa dla int/uint32_t, a nie long ://

+++ src/pubdir/pubdir.h
@@ -9,7 +9,7 @@ typedef struct {
 	char nickname[64];		// przezwisko
 	char email[128];		// email
 	char birthyear[4];		// rok
-	long gender;			// Płeć
+	int gender;			// Płeć
 	char city[64];			// Miejscowość
 	char phone[16];			// numer telefonu (wraz z nr kier.)
 } PubDir;
